+++ docs/history.rst
@@ -6,7 +6,10 @@ Latest
 
 3.5.0
 -----
-- ENH: Add `return_back_azimuth: bool` to allow compatibility between the azimuth output of `fwd` and `fwd_intermediate` (issue #1163)
+- ENH: Add `return_back_azimuth: bool` to allow compatibility between the azimuth output of the following functions (issue #1163):
+    `fwd` and `fwd_intermediate`, `inv` and `inv_intermediate`,
+    Note: BREAKING CHANGE for the default value `return_back_azimuth=True` in the functions `fwd_intermediate` and `inv_intermediate`
+    to mach the default value in `fwd` and `inv`
 
 3.4.1
 -----